net: tcp: Remove broken receive window computation
The receive window was always being filled as zero, and the
implementation of get_recv_wnd is awfully confusing as to what was
intended.

But this logic is needless anyway, the current architecture hands off
each received packet to a receive callback synchronously.  There is no
queueing inside the stack, so the window size never needs to change.
The only complexity is that there appear to be two existing tunables
that affect it, we have to pick a (compile time) minimum.

Note this also removes the recv_wnd from struct net_tcp, as it no
longer needs to be tracked.

Change-Id: I58c7b2753f4714f4751d64630ca7f09823b5a6a8
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>